requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
multidict==6.2.0
packaging==24.2
propcache==0.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-weather==2.0.7
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
yarl==1.18.3
